split up target window classes into their own files
