+++ Core/Battle/Menu/BattleMenus.cs
@@ -321,186 +321,6 @@ namespace OpenVIII
             return Victory_Menu.Update();
         }
 
-        #endregion Methods
-
-        #region Classes
-
-        public class IGMData_TargetGroup : IGMData_Group
-        {
-            public IGMData_TargetGroup(params IGMData[] d) : base(d)
-            {
-            }
-
-            public override bool Inputs()
-            {
-                IGMData_TargetEnemies i = (IGMData_TargetEnemies)(((IGMDataItem_IGMData)ITEM[0, 0]).Data);
-                IGMData_TargetParty ii = (IGMData_TargetParty)(((IGMDataItem_IGMData)ITEM[1, 0]).Data);
-                bool ret = false;
-                if (i.Enabled && (((i.Cursor_Status | ii.Cursor_Status) & Cursor_Status.Enabled) == 0 || !ii.Enabled))
-                    i.Cursor_Status |= Cursor_Status.Enabled;
-                else if (ii.Enabled && (((i.Cursor_Status | ii.Cursor_Status) & Cursor_Status.Enabled) == 0 || !i.Enabled))
-                    ii.Cursor_Status |= Cursor_Status.Enabled;
-                if (i.Enabled && (i.Cursor_Status & Cursor_Status.Enabled) != 0)
-                {
-                    ret = i.Inputs();
-                }
-                else if (ii.Enabled && (ii.Cursor_Status & Cursor_Status.Enabled) != 0)
-                {
-                    ret = ii.Inputs() || ret;
-                }
-                if(!ret)
-                {
-                    Cursor_Status = Cursor_Status.Hidden | Cursor_Status.Static | Cursor_Status.Enabled;
-                    skipdata = true;
-                    ret = base.Inputs();
-                    skipdata = false;
-                }
-                return ret;
-            }
-            public override bool Inputs_CANCEL()
-            {
-                Hide();
-                return base.Inputs_CANCEL();
-            }
-        }
-
-        public class IGMData_TargetEnemies : IGMData
-        {
-            #region Constructors
-
-            public IGMData_TargetEnemies(Rectangle pos) : base(6, 1, new IGMDataItem_Box(pos: pos, title: Icons.ID.TARGET), 2, 3)
-            {
-            }
-
-            #endregion Constructors
-
-            #region Methods
-
-            public override void Refresh()
-            {
-                if (Memory.State?.Characters != null)
-                {
-                    int pos = 0;
-                    if (Enemy.EnemyParty != null)
-                    {
-                        foreach (Enemy e in Enemy.EnemyParty)
-                        {
-                            //if(e.EII)
-                            ITEM[pos, 0] = new IGMDataItem_String(e.Name, SIZE[pos], Font.ColorID.White);
-                            ITEM[pos, 0].Show();
-                            BLANKS[pos] = false;
-
-                            pos++;
-                        }
-                        for (; pos < Count; pos++)
-                        {
-                            ITEM[pos, 0]?.Hide();
-                            BLANKS[pos] = true;
-                        }
-                    }
-                }
-            }
-
-            public override void Inputs_Left()
-            {
-                if (CURSOR_SELECT < 3)
-                {
-                    Cursor_Status &= ~Cursor_Status.Enabled;
-                    Menu.BattleMenus.Target_Party.Cursor_Status |= Cursor_Status.Enabled;
-                    Menu.BattleMenus.Target_Party.CURSOR_SELECT = CURSOR_SELECT % 3;
-                }
-                else
-                {
-                    SetCursor_select(CURSOR_SELECT - 3);
-                }
-                base.Inputs_Left();
-            }
-            public override void Inputs_Right()
-            {
-                if (CURSOR_SELECT >= 3 || (ITEM[3, 0] == null || !ITEM[3, 0].Enabled))
-                {
-                    Cursor_Status &= ~Cursor_Status.Enabled;
-                    Menu.BattleMenus.Target_Party.Cursor_Status |= Cursor_Status.Enabled;
-                    Menu.BattleMenus.Target_Party.CURSOR_SELECT = CURSOR_SELECT % 3;
-                }
-                else
-                {
-                    SetCursor_select(CURSOR_SELECT + 3);
-                }
-                base.Inputs_Right();
-            }
-
-            protected override void InitShift(int i, int col, int row)
-            {
-                base.InitShift(i, col, row);
-                SIZE[i].Inflate(-18, -20);
-                SIZE[i].Y -= 7 * row + 2;
-                //SIZE[i].Inflate(-22, -8);
-                //SIZE[i].Offset(0, 12 + (-8 * row));
-                SIZE[i].Height = (int)(12 * TextScale.Y);
-            }
-
-            #endregion Methods
-        }
-
-        public class IGMData_TargetParty : IGMData
-        {
-            #region Constructors
-
-            public IGMData_TargetParty(Rectangle pos) : base(3, 1, new IGMDataItem_Box(pos: pos, title: Icons.ID.NAME), 1, 3)
-            {
-            }
-
-            #endregion Constructors
-
-            #region Methods
-
-            public override void Refresh()
-            {
-                if (Memory.State?.Characters != null)
-                {
-                    List<KeyValuePair<int, Characters>> party = Memory.State.Party.Select((element, index) => new { element, index }).ToDictionary(m => m.index, m => m.element).Where(m => !m.Value.Equals(Characters.Blank)).ToList();
-                    byte pos = 0;
-                    foreach (KeyValuePair<int, Characters> pm in party)
-                    {
-                        Saves.CharacterData data = Memory.State[Memory.State.PartyData[pm.Key]];
-                        bool ded = data.IsDead;
-                        ITEM[pos, 0] = new IGMDataItem_String(Memory.Strings.GetName(pm.Value), SIZE[pos], ded ? Font.ColorID.Dark_Gray : Font.ColorID.White);
-
-                        pos++;
-                    }
-                }
-            }
-
-            protected override void InitShift(int i, int col, int row)
-            {
-                base.InitShift(i, col, row);
-                SIZE[i].Inflate(-18, -20);
-                SIZE[i].Y -= 7 * row + 2;
-                //SIZE[i].Inflate(-22, -8);
-                //SIZE[i].Offset(0, 12 + (-8 * row));
-                SIZE[i].Height = (int)(12 * TextScale.Y);
-            }
-
-            public override void Inputs_Left()
-            {
-                Cursor_Status &= ~Cursor_Status.Enabled;
-                Menu.BattleMenus.Target_Enemies.Cursor_Status |= Cursor_Status.Enabled;
-                Menu.BattleMenus.Target_Enemies.CURSOR_SELECT = (CURSOR_SELECT +3) % 6;
-                base.Inputs_Left();
-            }
-
-            public override void Inputs_Right()
-            {
-                Cursor_Status &= ~Cursor_Status.Enabled;
-                Menu.BattleMenus.Target_Enemies.Cursor_Status |= Cursor_Status.Enabled;
-                Menu.BattleMenus.Target_Enemies.CURSOR_SELECT = CURSOR_SELECT % 3;
-                base.Inputs_Right();
-            }
-
-            #endregion Methods
-        }
-
-        #endregion Classes
+#endregion Methods
     }
 }
